Refactor calculator component
