+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Calculator from './components/Calculator';
+import Calculator from './components/calculator/Calculator';
 
 class App extends React.Component {
   constructor() {
